Nft widget refactoring
