+++ src/widgets/nft-widget/components/index.ts
@@ -1,0 +1,2 @@
+export * from './nft-collection-view';
+export * from './nft-list-view';
